Ticket 684: "Analyze only" with bropo creates invalid DBZH datafield
The function PadNrays in ropo_realtime.py (accidentally?) modified the input scan, instead of just applying the change to the newscan-parameter that is returned from the function. This change caused the the DBZH datafield of the original scan to be modified. The only caused problems with the use of "analyze only", since the modified datafield was otherwise overwritten again with a corrected one.

The solve this, a clone of the datafield is used instead of the original field when making the modification.